Finished Video Playing functionality

+++ PeripheralTech.Mobile/PeripheralTech.Mobile/ViewModels/ProductVideoViewModel.cs
@@ -17,8 +17,14 @@ namespace PeripheralTech.Mobile.ViewModels
         public ProductVideoViewModel()
         {
             InitCommand = new Command(async () => await Init());
+            PlayPauseCommand = new Command(async () => await PlayPause());
+            //PauseCommand = new Command(async () => await Pause());
+            LeaveCommand = new Command(async () => await Leave());
         }
         public ICommand InitCommand { get; set; }
+        public ICommand PlayPauseCommand { get; set; }
+        //public ICommand PauseCommand { get; set; }
+        public ICommand LeaveCommand { get; set; }
 
         private ProductVideo productVideo;
         public ProductVideo ProductVideo
@@ -35,18 +41,51 @@ namespace PeripheralTech.Mobile.ViewModels
         public async Task Init()
         {
             ProductVideo = await _productVideoService.GetById<Model.ProductVideo>(ProductVideoID);
-
+        }
+        public async Task PlayPause()
+        {
             var video = CrossMediaManager.Current;
-            fileName = FileHelper.SaveFile(productVideo.Video, Guid.NewGuid() + ".mp4");
+            bool check = true;
+
             if (!string.IsNullOrEmpty(fileName))
+            {
+                await video.PlayPause();
+                check = false;
+            }
+
+            if (string.IsNullOrEmpty(fileName))
+            {
+                fileName = FileHelper.SaveFile(productVideo.Video, Guid.NewGuid() + ".mp4");
+            }
+            
+            if (!string.IsNullOrEmpty(fileName) && check)
             {
                 await video.Play(fileName);
 
                 video.MediaItemFinished += (sender, args) =>
                 {
                     FileHelper.DeleteFile(fileName);
+                    fileName = null;
                 };
             }
+        }
+        //public async Task Pause()
+        //{
+        //    var video = CrossMediaManager.Current;
+        //    if (video.IsPlaying())
+        //    {
+        //        await video.Pause();
+        //    }
+        //}
+        public async Task Leave()
+        {
+            var video = CrossMediaManager.Current;
+            if (video.IsPlaying())
+            {
+                FileHelper.DeleteFile(fileName);
+                await video.Stop();
+            }
+            await Application.Current.MainPage.Navigation.PopModalAsync();
         }
     }
 }
